backend/requirements.txt
python-dotenv==1.0.1
fastapi==0.115.5
llama_index==0.12.0
numpy==1.26.4
openai==1.55.0
pydantic==2.9.2
pymongo==4.9.2
uvicorn==0.32.0
langchain==0.3.7
scikit-learn==1.5.2
langchain_groq==0.2.1
passlib==1.7.4
motor==3.6.0
bcrypt==3.2.0
async_timeout==4.0.3